feat: [Wasm] Adjust netcore compatiblity
